[Refactor] 최초 공정표 문서 등록 , 전체 공정표 화면 컴포넌트 분리
